rename "adhoc" sample to "connect"
