+++ sdk/cosmosdb/azure_data_cosmos/examples/cosmos_connect.rs
@@ -1,0 +1,47 @@
+// REVIEW: Don't merge this file. It's just for adhoc testing purposes.
+
+use azure_data_cosmos::{CosmosClient, CosmosClientMethods, DatabaseClientMethods};
+use clap::Parser;
+
+/// A simple example to show connecting to a Cosmos Account and retrieving the properties of a database.
+#[derive(Parser)]
+pub struct Args {
+    /// The CosmosDB endpoint to connect to.
+    endpoint: String,
+
+    /// The database to fetch information for.
+    database: String,
+
+    /// An authentication key to use when connecting to the Cosmos DB account. If omitted, the connection will use Entra ID.
+    #[clap(long)]
+    #[cfg(feature = "key-auth")]
+    key: Option<String>,
+}
+
+#[tokio::main]
+pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let args = Args::parse();
+
+    let client = create_client(&args);
+
+    let db_client = client.database(&args.database);
+    let response = db_client.read(None).await?.deserialize_body().await?;
+    println!("{:?}", response);
+    Ok(())
+}
+
+#[cfg(feature = "key-auth")]
+fn create_client(args: &Args) -> CosmosClient {
+    if let Some(key) = args.key.as_ref() {
+        CosmosClient::with_shared_key(&args.endpoint, key.clone(), None).unwrap()
+    } else {
+        let cred = azure_identity::create_default_credential().unwrap();
+        CosmosClient::new(&args.endpoint, cred, None).unwrap()
+    }
+}
+
+#[cfg(not(feature = "key-auth"))]
+fn create_client(args: &Args) -> CosmosClient {
+    let cred = azure_identity::create_default_credential().unwrap();
+    CosmosClient::new(&args.endpoint, cred, None).unwrap()
+}
